+++ js/src/forum/components/ListItem.js
@@ -116,6 +116,13 @@ export default class listItem extends Component {
     /* Jump to the last relevant post (first unread or last post) */
     const jumpTo = Math.min(discussion.lastPostNumber() ?? 0, (discussion.lastReadPostNumber() || 0) + 1);
 
+    /* setting post counts & text */
+		console.log("unread count: " + discussion.unreadCount());
+		const replyText = discussion.unreadCount() 
+			? app.translator.trans("walsgit_discussion_cards.forum.unreadReplies", { count: discussion.unreadCount()} ) 
+			: app.translator.trans("walsgit_discussion_cards.forum.replies", { count: discussion.replyCount() || "0"} );
+    const postCount = discussion.unreadCount() ? discussion.unreadCount() + "*" : discussion.replyCount();
+
     return (
       <div key={discussion.id()}
           data-id={discussion.id()}
@@ -174,7 +181,7 @@ export default class listItem extends Component {
                   </h2>
                   {app.screen() !== 'phone' && Number(settings.showReplies) === 1 && Number(settings.showRepliesOnRight) === 1 ?
                   <div className="DiscussionListItem-count">
-                    <span aria-hidden="true">{abbreviateNumber(discussion.replyCount())}</span>
+                    <span aria-hidden="true">{abbreviateNumber(postCount)}</span>
 
                     <span className="visually-hidden">
                       {app.translator.trans('core.forum.discussion_list.unread_replies_a11y_label', { count: discussion.replyCount() })}
@@ -214,7 +221,7 @@ export default class listItem extends Component {
                         {m(LastReplies, {discussion: discussion})}
                       </div>
                       <div className="Repcount">
-                        {app.translator.trans('walsgit_discussion_cards.forum.replies', {count: discussion.replyCount() || '0'})}
+                        {replyText}
                       </div>
                     </div>
                     <div className="Arrow">
@@ -225,7 +232,7 @@ export default class listItem extends Component {
                 : Number(settings.showReplies) === 1 && !Number(settings.showRepliesOnRight) ?
                   <div className="imageLabel discussionReplyCount">
                     {icon('fas fa-comment', {className: 'labelIcon'})}
-                    {discussion.replyCount()}
+                    {postCount}
                   </div> : ''
               }
             </div>